ci: remove wrong/unuseful 'branches_ignore' and 'tags' entries and add 'workflow_dispatch'
The 'renovate/**' 'branches_ignore' entry prevented renovate pushes to trigger ci that is needed because PR is skipped as per the 'dispatcher' step. The other entries are unuseful for our workflow. Add the possibility to trigger the workflow manually. By default with no 'branches' and 'tags' filters the workflow will run for events affecting either branches or tags.